factory function to create rule
